refactor: MoveStrategy 패키지 이동
