+++ src/main/kotlin/racingcar/domain/movestrategy/MoveStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.movestrategy
 
 interface MoveStrategy {
