+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -2,6 +2,7 @@ package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.domain.movestrategy.MoveStrategy
 
 internal class CarTest {
